--- packages/furo-navigation/test/furo-flat-tree_test.html
@@ -1,103 +1,0 @@
-<!doctype html>
-
-<html>
-<head>
-  <title>test for furo-tree</title>
-  <meta charset="utf-8">
-  <meta name="viewport" content="width=device-width, minimum-scale=1.0, initial-scale=1, user-scalable=yes">
-
-  <meta name="viewport" content="width=device-width, minimum-scale=1.0, initial-scale=1, user-scalable=yes">
-  <script src="/node_modules/@webcomponents/webcomponentsjs/webcomponents-loader.js"></script>
-  <script src="/node_modules/mocha/mocha.js"></script>
-  <script src="/node_modules/chai/chai.js"></script>
-  <script src="/node_modules/@polymer/test-fixture/test-fixture.js"></script>
-  <script src="/node_modules/wct-mocha/wct-mocha.js"></script>
-  <script src="/node_modules/sinon/pkg/sinon.js"></script>
-
-  <script type="module">
-
-    import "../furo-catalog.js"
-    import "../furo-flat-tree"
-    import "./produce-data"
-    import "./initEnv"
-    import "@furo/data"
-
-    import "@furo/fbp/flow-bind"
-  </script>
-</head>
-<body>
-
-<test-fixture id="basic">
-  <template>
-    <flow-bind>
-      <template>
-        <produce-data @-data="--data"></produce-data>
-        <entity-object type="vnd.com.acme.tree" ƒ-inject-raw="--data" @-object-ready="--entityObj"></entity-object>
-        <furo-flat-tree ƒ-bind-data="--entityObj" ></furo-flat-tree>
-
-      </template>
-    </flow-bind>
-  </template>
-</test-fixture>
-
-<script type="module">
-
-  import '/node_modules/axe-core/axe.min.js';
-  import {axeReport} from '/node_modules/pwa-helpers/axe-report.js';
-
-  describe('furo-tree tests ', function () {
-
-    let tree, entity, dataprocuder, host;
-    beforeEach(async () => {
-      let fix = fixture('basic');
-      host = fix[0]._host;
-      entity = fix[2];
-      await entity.updateComplete;
-      dataprocuder = fix[1];
-      await dataprocuder.updateComplete;
-      tree = fix[3];
-      await tree.updateComplete;
-
-      tree.__tmp = tree._buildFlatTree;
-      tree._buildFlatTree = (node)=>{
-        tree.__tmp(node);
-        /**
-         * @event tree-parsed
-         * Fired when
-         * detail payload:
-         */
-        let customEvent = new Event('tree-parsed', {composed:true, bubbles: true});
-        customEvent.detail = this._flatTree;
-        tree.dispatchEvent(customEvent)
-
-      }
-    });
-
-
-    it('should build flat tree', (done) => {
-      tree.addEventListener("tree-parsed",(e)=>{
-        assert.equal(tree._flatTree.length,9);
-        done();
-      });
-      dataprocuder.produce();
-
-
-    });
-
-
-    // a11y tests
-    it('a11y', () => {
-      return axeReport(fixture('basic'));
-    });
-
-    it('should be a furo-tree', (done) => {
-      assert.equal(tree.nodeName, "FURO-FLAT-TREE");
-      done()
-    });
-
-
-  })
-  ;
-</script>
-</body>
-</html>
