+++ packages/furo-navigation/test/index.html
@@ -17,6 +17,7 @@
 <script>
   // Load and run all tests (.html, .js):
   WCT.loadSuites([
+      'furo-tree_test.html'
   ]);
 </script>
 </body>
