+++ wct.conf.json
@@ -4,7 +4,8 @@
         "node_modules/@furo/form/test/index.html",
         "node_modules/@furo/route/test/index.html",
         "node_modules/@furo/data/test/index.html",
-        "node_modules/@furo/util/test/index.html"
+        "node_modules/@furo/util/test/index.html",
+        "node_modules/@furo/navigation/test/index.html"
     ],
     "root": "./",
     "npm": true,
@@ -19,6 +20,7 @@
                 "node_modules/@furo/fbp/*.js",
                 "node_modules/@furo/route/*.js",
                 "node_modules/@furo/form/*.js",
+                "node_modules/@furo/navigation/*.js",
                 "node_modules/@furo/util/*.js"
             ],
             "exclude": ["**/furo-catalog.js"]
